Shift components to use Next's path router instead of the hash router
