+++ packages/app/pages/manage.tsx
@@ -10,14 +10,15 @@ import {
 } from "@cabindao/topo";
 import { useEffect, useMemo, useState } from "react";
 import type { Contract, ContractSendMethod } from "web3-eth-contract";
-import { contractAddressesByNetworkId, getAbiFromJson } from "./constants";
-import { useAddress, useChainId, useWeb3 } from "./Web3Context";
+import { contractAddressesByNetworkId, getAbiFromJson } from "../components/constants";
+import { useAddress, useChainId, useWeb3 } from "../components/Web3Context";
 import passportJson from "@cabindao/nft-passport-contracts/artifacts/contracts/Passport.sol/Passport.json";
 import passportFactoryJson from "@cabindao/nft-passport-contracts/artifacts/contracts/PassportFactory.sol/PassportFactory.json";
 import { PlusIcon, MinusIcon } from "@radix-ui/react-icons";
 import { TransactionReceipt } from "web3-core";
 import Papa from "papaparse";
-import { resolveAddress } from "./utils";
+import { resolveAddress } from "../components/utils";
+import Layout from "../components/Layout";
 
 const SmallBox = styled(Box, {
   width: "25%",
@@ -371,4 +372,12 @@ const ManageTabContent = () => {
   );
 };
 
-export default ManageTabContent;
+const ManagePage = () => {
+  return (
+    <Layout>
+      <ManageTabContent />
+    </Layout>
+  )
+}
+
+export default ManagePage;
